more cleanup of Jira connector
